requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
comm==0.2.3
comtypes==1.4.14
confluent-kafka==2.12.2
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
distro==1.9.0
executing==2.2.1
faiss-cpu==1.12.0
filelock==3.20.1
fonttools==4.60.1
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-genai==1.56.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
groq==1.0.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
kafka-python==2.3.0
kiwisolver==1.4.9
langchain==1.0.0
langchain-core==1.0.0
langgraph==1.0.0
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.0
langgraph-sdk==0.2.9
langsmith==0.4.37
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
MouseInfo==0.1.3
mpmath==1.3.0
narwhals==2.10.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numpy==2.2.6
openai==2.5.0
opencv-python==4.12.0.88
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
patsy==1.0.2
pillow==11.0.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==5.29.1
psutil==7.2.1
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pyparsing==3.2.5
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pyspark==3.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytweening==1.2.0
pytz==2025.2
pywin32==311
pywinauto==0.6.9
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.28.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
statsmodels==0.14.6
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
toml==0.10.2
torch==2.9.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uiautomation==2.0.29
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.14
websockets==15.0.1
wheel==0.45.1
xxhash==3.6.0
zstandard==0.25.0
